requirements.txt

django-serializer==0.0.8

xhtml2pdf==0.2.5

psycopg2-binary==2.8.6
djangorestframework==3.12.4
Django==3.1.4
gunicorn==20.0.4
twilio==6.53.0

django-cors-headers==3.7.0

django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-sendsms==0.4
django-twilio==0.13.1.post0


email-validator==1.1.2

decorator==4.4.2
html5lib==1.1
httplib2==0.14.0